+++ fabric/src/main/java/band/kessokuteatime/bounced/fabric/BouncedFabric.java
@@ -1,0 +1,9 @@
+package band.kessokuteatime.bounced.fabric;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public final class BouncedFabric implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+	}
+}