+++ neoforge/src/main/java/band/kessokuteatime/bounced/neoforge/BouncedNeoForge.java
@@ -1,0 +1,9 @@
+package band.kessokuteatime.bounced.neoforge;
+
+import band.kessokuteatime.bounced.Bounced;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.fml.common.Mod;
+
+@Mod(value = Bounced.ID, dist = Dist.CLIENT)
+public final class BouncedNeoForge {
+}